Add an Enemy node class
Add an Enemy class that inherits from Area3D. It handles being hitting and being hit by the player. It can also show which objects it's currently colliding with.
The test enemy now moves in the direction that the arrow was going.
Also made the arrow delete itself when it hits a wall or goes to far from the center
Also formatted the code to match the Godot style guide