Change the Orc chasing logic. Add some animations to the Orc
Change the Orc chasing logic so that it goes to where the player was half a
second ago.

Add animations to the Orc for when it moves. I still need to add the animations
for attacking the player. I also still need to add animations to the player for
when it gets hit