Add health and damages to Enemy and Player. Add attack animation to Enemy
Added health values to Enemy and Player. Added damge values to the Enemy and
each one of the Player's attacks. The Enemy will take damage when it gets hit by
the player. I still need to make the Arrow deal damage based on its speed.
I also still need to make the Player take damage.

Add animations to the Enemy for attacking the Player